проверяет авторизацию в компоненте authorization-nav
